Rename max_pool_2d to pool_2d and Pool for the DownsampleFactorMax
